feat : swagger linux version

+++ src/server/src/main/java/com/econovation/recruit/domain/interviewer/InterviewerRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InterviewerRepository extends JpaRepository<Interviewer, Long> {
+public interface InterviewerRepository extends JpaRepository<Interviewer, Integer> {
 }
